Moved libraries to lib directory
